pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tinypg"
version = "0.2.0"
description = "Ephemeral PostgreSQL databases for Python development and testing"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "kketch", email = "9odxnqhjr@mozmail.com"}
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Testing",
]
requires-python = ">=3.8"
dependencies = [
    "psycopg2-binary>=2.9.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
]
async = [
    "asyncpg>=0.29.0",
]
docs = [
    "pdoc>=14.4",
]

[project.urls]
Homepage = "https://github.com/kketch/tinypg"
Documentation = "https://python-tinypg.readthedocs.io"
Repository = "https://github.com/kketch/tinypg"

[tool.setuptools.packages.find]
where = ["src"]

# [tool.setuptools_scm]
# write_to = "src/tinypg/_version.py"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pdoc]
modules = ["tinypg"]
docformat = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
]
